moving getPaths, etc. to FileTable.
