+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/csv/CSVDataSourceV2.scala
@@ -30,16 +30,10 @@ class CSVDataSourceV2 extends FileDataSourceV2 {
   override def shortName(): String = "csv"
 
   override def getTable(options: CaseInsensitiveStringMap): Table = {
-    val paths = getPaths(options)
-    val tableName = getTableName(options, paths)
-    val optionsWithoutPaths = getOptionsWithoutPaths(options)
-    CSVTable(tableName, sparkSession, optionsWithoutPaths, paths, None, fallbackFileFormat)
+    CSVTable(sparkSession, this, options, None, fallbackFileFormat)
   }
 
   override def getTable(options: CaseInsensitiveStringMap, schema: StructType): Table = {
-    val paths = getPaths(options)
-    val tableName = getTableName(options, paths)
-    val optionsWithoutPaths = getOptionsWithoutPaths(options)
-    CSVTable(tableName, sparkSession, optionsWithoutPaths, paths, Some(schema), fallbackFileFormat)
+    CSVTable(sparkSession, this, options, Some(schema), fallbackFileFormat)
   }
 }
