+++ sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/v2/json/JsonDataSourceV2.scala
@@ -30,17 +30,11 @@ class JsonDataSourceV2 extends FileDataSourceV2 {
   override def shortName(): String = "json"
 
   override def getTable(options: CaseInsensitiveStringMap): Table = {
-    val paths = getPaths(options)
-    val tableName = getTableName(options, paths)
-    val optionsWithoutPaths = getOptionsWithoutPaths(options)
-    JsonTable(tableName, sparkSession, optionsWithoutPaths, paths, None, fallbackFileFormat)
+    JsonTable(sparkSession, this, options, None, fallbackFileFormat)
   }
 
   override def getTable(options: CaseInsensitiveStringMap, schema: StructType): Table = {
-    val paths = getPaths(options)
-    val tableName = getTableName(options, paths)
-    val optionsWithoutPaths = getOptionsWithoutPaths(options)
-    JsonTable(tableName, sparkSession, optionsWithoutPaths, paths, Some(schema), fallbackFileFormat)
+    JsonTable(sparkSession, this, options, Some(schema), fallbackFileFormat)
   }
 }
